backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.15.0,<0.16.0
python-dotenv>=0.19.0
httpx>=0.24.0
pydantic>=2.7.0
redis>=4.0.0
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.0.1
openai>=1.0.0
aiohttp>=3.8.0
web3>=6.0.0  # For on-chain token lookups
agno>=1.5.4  # For portfolio analysis
anthropic>=0.18.0  # Required by agno for Claude integration
duckduckgo-search>=4.1.1  # Required by agno for web search
yfinance>=0.2.36  # Required by agno for financial data
lancedb>=0.5.0  # Required by agno for vector storage
tantivy>=0.21.0  # Required by agno for search
pypdf>=4.0.0  # Required by agno for document processing
exa-py>=1.0.0  # Required by agno for enhanced analysis
newspaper4k>=0.1.0  # Required by agno for web scraping
sqlalchemy>=2.0.0  # Required by agno for database operations
firecrawl-py>=0.0.16  # For web scraping DeFi protocol data
